Documentation: Receiver files.

+++ src/receiver/pipeline-stages/SignalStretcher.h
@@ -3,12 +3,26 @@
 
 #include<inttypes.h>
 
+/**
+ * @brief A class the implements horizontal stretching of a signal using
+ *      Linear Interpolation. 
+ * 
+ */
 class SignalStretcher {
 
     public:
         SignalStretcher() {}
 
-        void StretchSignal(uint16_t * signal, int sLength, double * dest, int dLength) {
+        /**
+         * @brief Stretch a signal from length \p sLength to length \p dLength using
+         *      Linear Interpolation. Works with integer input data.
+         * 
+         * @param signal    - Signal data - integer type.
+         * @param sLength   - Signal length.
+         * @param dest      - Output buffer. Has to be of length at least \p dLength .
+         * @param dLength   - Output length to stretch up/down to.
+         */
+        void StretchSignal(uint16_t * signal, int sLength, float * dest, int dLength) {
 
             int i = 0;
             float coeff = (float)(sLength-1) / (dLength-1);
@@ -26,6 +40,15 @@ class SignalStretcher {
 
         }
     
+        /**
+         * @brief Stretch a signal from length \p sLength to length \p dLength using
+         *      Linear Interpolation. Works with float input data.
+         * 
+         * @param signal    - Signal data - float type.
+         * @param sLength   - Signal length.
+         * @param dest      - Output buffer. Has to be of length at least \p dLength .
+         * @param dLength   - Output length to stretch up/down to.
+         */
         void StretchSignal(float * signal, int sLength, float * dest, int dLength) {
 
             int i = 0;
